♻️ [refactor 전략 엔티티 타입 수정 및 주석 변경

+++ src/main/java/com/investmetic/domain/strategy/model/entity/StrategyStatistics.java
@@ -8,7 +8,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
-import java.lang.Long;
 import java.time.LocalDate;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -30,7 +29,7 @@ public class StrategyStatistics extends BaseEntity {
 
     private String operationPeriod; // 운용 기간
 
-    private Long cumulativeTransactionAmount; // 누적 거래금액
+    private Long cumulativeTransactionAmount; // 누적 입출금액
 
     private LocalDate startDate; // 시작일
 
@@ -42,37 +41,37 @@ public class StrategyStatistics extends BaseEntity {
 
     private Long cumulativeProfitAmount; // 누적 수익금액
 
-    private Long cumulativeProfitRate; // 누적 수익률
+    private Double cumulativeProfitRate; // 누적 수익률
 
-    private Long recentYearProfitRate; // 최근 1년 수익률
+    private Double recentYearProfitRate; // 최근 1년 수익률
 
     private Long maxCumulativeProfitAmount; // 최대 누적 수익금액
 
-    private Long maxCumulativeProfitRate; // 최대 누적 수익률
+    private Double maxCumulativeProfitRate; // 최대 누적 수익률
 
     private Long averageProfitLossAmount; // 평균 손익 금액
 
-    private Long averageProfitLossRate; // 평균 손익률
+    private Double averageProfitLossRate; // 평균 손익률
 
-    private Long maxDailyProfitAmount; // 최대 일간 수익금액
+    private Long maxDailyProfitAmount; // 최대 일수익 금액
 
-    private Long maxDailyProfitRate; // 최대 일간 수익률
+    private Double maxDailyProfitRate; // 최대 일 수익률
 
-    private Long maxDailyLossAmount; // 최대 일간 손실금액
+    private Long maxDailyLossAmount; // 최대 일손실 금액
 
-    private Long maxDailyLossRate; // 최대 일간 손실률
+    private Double maxDailyLossRate; // 최대 일 손실률
 
-    private Long roa; // ROA
+    private Double roa; // ROA
 
-    private Long profitFactor; // Profit Factor
+    private Double profitFactor; // Profit Factor
 
-    private Long currentDrawdown; // 현재 낙폭
+    private Long currentDrawdown; // 현재 자본 인하 금액
 
-    private Long currentDrawdownRate; // 현재 낙폭률
+    private Double currentDrawdownRate; // 현재 자본 인하율
 
-    private Long maxDrawdown; // 최대 낙폭
+    private Long maxDrawdown; // 최대 자본 인하 금액
 
-    private Long maxDrawdownRate; // 최대 낙폭률
+    private Double maxDrawdownRate; // 최대 자본 인하율
 
     private Integer currentConsecutiveProfitLossDays; // 현재 연속 손익일수
 
@@ -86,13 +85,13 @@ public class StrategyStatistics extends BaseEntity {
 
     private Double winRate; // 승률
 
-    private Long smScore; // SM Score
+    private Double smScore; // SM Score
 
-    private Long kpRatio; // KP Ratio
+    private Double kpRatio; // KP Ratio
 
     private Long initialInvestment; // 최초 투자금액
 
     private LocalDate finalProfitLossDate; // 최종 손익 일자
 
-    private Integer totalTradeDays; // 총 거래일수
+    private Integer totalTradeDays; // 총 매매일수
 }
